+++ src/components/app/Page.svelte
@@ -108,10 +108,4 @@
         gap: var(--wordplay-spacing);
         background: var(--wordplay-background);
     }
-
-    .fullscreen:not(:hover) {
-        opacity: 0.2;
-        border: none;
-        background: none;
-    }
 </style>